sdl-build: Check if a suitable SDL2 version is available locally before fetching it
CMake offers the find_package functionality which can be automatically called by
FetchContent_Declare. We are doing so to see if a suitable SDL2 version is locally
installed. SDL2-2.30.0 will be available with the upcoming ubuntu 24.04 (Noble) and
has also made its way into ArchLinux on 5th of february 2024.
We can therefore expect that a reasonable amount of users do not have to build SDL2
but can use the system version.

Signed-off-by: Mimoja <git@mimoja.de>