sdl-build: Actually configure SDL to not build a shared library
The SDL build configuration is not respected in certain setups as the declaration
is never passed to the SDL cmake.

Due to changes within from cmake 3.13+ it is possible to set SDL_{SHARED,STATIC,TEST}
directly as also acknoledged by the SDL team:
https://github.com/libsdl-org/SDL/commit/803924b2233458e4a2a7d24530e2208940ca278b
The change is availabel in SDL2-2.25.1+ which is well below our SDL2 target.
It is also worth noting that even the old Debian 11 ships a comptablie
cmake version 3.25.1
We therefore opt to use the shorter and more ledigable parameters over the
_ENABLED_BY_DEFAULT sibilings.

Signed-off-by: Mimoja <git@mimoja.de>